Rename collections and contact us services. Add app and env params to all their requests. Remove unused CountriesService
